Centered all charge meters in minmode
Moved many hud elements to not collide with the new placement.

Added a workaround for (most) meters  to place the charge bar label inside of the bar itself.

Existing charge bars in the center of the screen have been tweaked for consistency.

Note that the scout meter and generic meter still have bugs to iron out, but should still be usable in minmode (they just wont look right)