Fixed playerlists going off the scoreboard
This only occurs if playing on a server or gamemode where a team has more than 12 players. Since the scrollbar tends to block player ping, I avoided implementing this at first.

If you regularly play on servers/gamemodes with teams of more than 12 players, swap the values on lines 228 and 267 to reduce the amount of times the scrollbar gets in your way.